Add files via upload
The vice president sends an Excel template containing BigTechCompany's quarterly financial statements and a PowerPoint template to complete. The task is to prepare tables on the company's financials and to present your analysis in a PowerPoint presentation. 

Some of the financial metrics you will be digging into include revenue, operating income (EBITDA), net income, and free cash flow. Be asked to calculate financial metrics, such as EBITDA margin and debt/EBITDA. If you aren't sure what this means, don't worry; your vice president included instructions and hints in the spreadsheet! 

Once you've finished your Excel analysis, you will use this information to draw some conclusions about the company. What trends do you see in the company's financials? Is the company stable, or is it in decline? Are there any red flags that would make this company high risk? 

These are important questions, as your team does not want to build a relationship with a client that is not sustainable or may run into severe financial problems in the future. Rather, your team wants to work with companies with stable earnings, at least $1 billion in annual earnings, and positive net income (profitability) or a path to positive net income in the future. Your bank wants to avoid companies with negative free cash flow and unhealthy leverage ratios.