Add files via upload
 To complete this task, presents your finding through a PPT:

Copy each of the tables into the PowerPoint template. There are stickers on each PowerPoint slide to help guide you through this process. 
On slide 3 of the template, write bullets on the company's customer trends. 
On slide 5, compose bullets on the company and whether or not you would recommend the bank work with this client. 
Once you've completed the presentation, submit your file below.